Archivos formateados, preparados para probar

+++ Curso/libft/ft_calloc.c
@@ -5,17 +5,17 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: dcerrato <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/20 12:01:09 by dcerrato          #+#    #+#             */
-/*   Updated: 2022/06/20 12:25:36 by dcerrato         ###   ########.fr       */
+/*   Created: 2022/06/20 13:44:23 by dcerrato          #+#    #+#             */
+/*   Updated: 2022/06/20 14:10:05 by dcerrato         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-void *ft_calloc(size_t count, size_t size)
+void	*ft_calloc(int count, size_t size)
 {
 	void	*mem;
-	int		i;
+	size_t	i;
 
 	if (count <= 0 || size <= 0)
 		return (NULL);
