+++ Curso/libft/ft_memcmp.c
@@ -6,17 +6,19 @@
 /*   By: dcerrato <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/17 17:23:57 by dcerrato          #+#    #+#             */
-/*   Updated: 2022/06/20 11:59:07 by dcerrato         ###   ########.fr       */
+/*   Updated: 2022/06/20 14:16:47 by dcerrato         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_memcmp(const void *str1, const void *str2, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*s1;
 	unsigned char	*s2;
 
-	if (str1 == NULL || str2 == NULL || n <= 0)
+	if (str1 == NULL || str2 == NULL || n <= 0)
 	{
 		if (str1 == NULL && str2 != NULL)
 			return (-1);
