+++ Curso/libft/ft_memcpy.c
@@ -6,13 +6,15 @@
 /*   By: dcerrato <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/17 12:21:00 by dcerrato          #+#    #+#             */
-/*   Updated: 2022/06/17 15:31:26 by dcerrato         ###   ########.fr       */
+/*   Updated: 2022/06/20 14:17:11 by dcerrato         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 void	*ft_memcpy(void *dst, const void *src, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	if (dst == NULL || src == NULL)
 		return (dst);
